+++ rest/web_employer.go
@@ -28,6 +28,10 @@ func postEmployer(c *gin.Context, db *redis.Client) {
 
 	newInitEmp := []rdb.Employee{phEmp}
 	newInitPay := []rdb.Payment{phPay}
+	newIdPw := rdb.Login{
+		Username: newEmployer,
+		Password: newPassword,
+	}
 	// permenent employee data table
 	_ = rdb.JsonSet(
 		db,
@@ -49,11 +53,56 @@ func postEmployer(c *gin.Context, db *redis.Client) {
 		HISTORICAL_PAYMENT_PATH,
 		&newInitPay,
 	)
+	// add employee id pw to login table
+	_ = rdb.JsonArrAppend(
+		db,
+		LOGIN_KEY,
+		LOGIN_PATH,
+		&newIdPw,
+	)
 	c.JSON(http.StatusOK, gin.H{
 		"message":    DATABASE_CREATE_SUCCESS,
 		"employerId": newEmployer,
 		"employerPw": newPassword,
 	})
+}
+
+func loginEmployer(c *gin.Context, db *redis.Client) {
+	username, err := handleParamEmployerId(c)
+	if err != nil {
+		return
+	}
+	password, err := handleParamEmployerPw(c)
+	if err != nil {
+		return
+	}
+
+	var registered [][]rdb.Login
+	_ = rdb.JsonGet(
+		db,
+		LOGIN_KEY,
+		LOGIN_PATH,
+		&registered,
+	)
+	id, pw := isRegistered(registered[0], username, password)
+	switch {
+	case id && pw:
+		c.JSON(http.StatusOK, gin.H{
+			"message":    LOGIN_SUCCESS,
+			"employerId": username,
+			"employerPw": password,
+		})
+	case id:
+		// if wrong password StatusUnauthorized
+		c.JSON(http.StatusUnauthorized, gin.H{
+			"message": LOGIN_ERROR + " wrong password",
+		})
+	default:
+		// if no id StatusForbidden
+		c.JSON(http.StatusForbidden, gin.H{
+			"message": LOGIN_ERROR + " id unrecognized",
+		})
+	}
 }
 
 func delEmployer(c *gin.Context, db *redis.Client) {